Split responsibilities of index file in two files
The index file was responsible for starting the importer and being an exportable
module to create import pipeline. These responsibilities are now in separate files
