+++ stream/importPipeline.js
@@ -1,0 +1,40 @@
+var spy = require('through2-spy');
+var logger = require('pelias-logger').get('openstreetmap-points');
+var categoryDefaults = require('../config/category_map');
+
+var streams = {};
+
+streams.config = {
+  categoryDefaults: categoryDefaults
+};
+
+streams.pbfParser = require('./pbf').parser;
+streams.docConstructor = require('./document_constructor');
+streams.docDenormalizer = require('./denormalizer');
+streams.tagMapper = require('./tag_mapper');
+streams.adminLookup = require('./adminLookup');
+streams.addressExtractor = require('./address_extractor');
+streams.deduper = require('./deduper');
+streams.categoryMapper = require('./category_mapper');
+streams.dbMapper = require('pelias-model').createDocumentMapperStream;
+streams.elasticsearch = require('pelias-dbclient');
+
+// default import pipeline
+streams.import = function(opts){
+  streams.pbfParser(opts)
+    .pipe( streams.docConstructor() )
+    .pipe( streams.tagMapper() )
+    .pipe( streams.docDenormalizer() )
+    .pipe( streams.addressExtractor() )
+    .pipe( streams.categoryMapper( categoryDefaults ) )
+    .pipe( streams.adminLookup() )
+    .pipe( streams.deduper() )
+    .pipe( spy.obj(function (doc) {
+        logger.info(doc.getGid(), doc.getName('default'), doc.getCentroid());
+      })
+    )
+    .pipe( streams.dbMapper() )
+    .pipe( streams.elasticsearch() );
+};
+
+module.exports = streams;
